+++ encoding/protobuf/error_test.go
@@ -18,10 +18,33 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package encoding
+package protobuf
 
-// StreamOption is an option that may be passed in at streaming function call
-// sites.
-type StreamOption interface {
-	unimplemented()
+import (
+	"errors"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/yarpc/yarpcerrors"
+)
+
+func TestNewOK(t *testing.T) {
+	err := NewError(yarpcerrors.CodeOK, "okay")
+	assert.Nil(t, err)
+
+	assert.Equal(t, yarpcerrors.FromError(err).Code(), yarpcerrors.CodeOK)
+	assert.Equal(t, yarpcerrors.FromError(err).Message(), "")
+}
+
+func TestNew(t *testing.T) {
+	err := NewError(yarpcerrors.CodeNotFound, "unfounded accusation")
+	assert.Equal(t, yarpcerrors.FromError(err).Code(), yarpcerrors.CodeNotFound)
+	assert.Equal(t, yarpcerrors.FromError(err).Message(), "unfounded accusation")
+	assert.Contains(t, err.Error(), "unfounded accusation")
+}
+
+func TestForeignError(t *testing.T) {
+	err := errors.New("to err is go")
+	assert.Equal(t, yarpcerrors.FromError(err).Code(), yarpcerrors.CodeUnknown)
+	assert.Equal(t, yarpcerrors.FromError(err).Message(), "to err is go")
 }