+++ internal/observability/call.go
@@ -113,8 +113,13 @@ func (c call) endStats(elapsed time.Duration, err error, isApplicationError bool
 		c.edge.latencies.Observe(elapsed)
 		return
 	}
-	// For now, assume that all application errors are the caller's fault.
-	if isApplicationError {
+
+	isStatus := yarpcerrors.IsStatus(err)
+
+	// If the error is not a yarpcerrors.Status, we cannot determine a code or
+	// fault. Assume that these application errors are the caller's fault and emit
+	// a generic error tag.
+	if isApplicationError && !isStatus {
 		c.edge.callerErrLatencies.Observe(elapsed)
 		if counter, err := c.edge.callerFailures.Get(_error, "application_error"); err == nil {
 			counter.Inc()
@@ -122,7 +127,7 @@ func (c call) endStats(elapsed time.Duration, err error, isApplicationError bool
 		return
 	}
 
-	if !yarpcerrors.IsStatus(err) {
+	if !isStatus {
 		c.edge.serverErrLatencies.Observe(elapsed)
 		if counter, err := c.edge.serverFailures.Get(_error, "unknown_internal_yarpc"); err == nil {
 			counter.Inc()
@@ -130,6 +135,7 @@ func (c call) endStats(elapsed time.Duration, err error, isApplicationError bool
 		return
 	}
 
+	// Emit finer grained metrics since the error is a yarpcerrors.Status.
 	errCode := yarpcerrors.FromError(err).Code()
 	switch errCode {
 	case yarpcerrors.CodeCancelled,